requirements.txt
gunicorn==19.4.5
Django==1.9.5
psycopg2==2.6.1
whitenoise==3.0
Pillow==3.2.0
django-classbasedsettings==1.5.0
python-coveralls==2.7.0
geopy==1.11.0
pinax-stripe==3.1.0
